11/19 - 1
내역 통신 일단 완료

+++ CheckMoney/app/src/main/java/com/checkmoney/account/CalTotal.kt
@@ -1,5 +1,5 @@
 package com.checkmoney.account
 
 interface CalTotal {
-    fun calTotal(deleteConsum: Int, deletePrice: Long, addConsum: Int, addPrice: Long)
+    fun calTotal(deleteConsum: Int, deletePrice: Int, addConsum: Int, addPrice: Int)
 }
